Simplify unsubscribe header and improve detection mechanism
