+++ app/internal_packages/list-unsubscribe/styles/unsubscribe.less
@@ -1,0 +1,10 @@
+@import 'ui-variables';
+
+.unsubscribe-action {
+  color: @text-color-very-subtle;
+  text-decoration: underline;
+
+  &:hover {
+    color: @text-color-subtle;
+  }
+}
