--- app/internal_packages/unsubscribe/styles/unsubscribe.less
@@ -1,95 +1,0 @@
-@import 'ui-variables';
-@import 'ui-mixins';
-
-.unsubscribe-wrapper {
-  cursor: default;
-  position: relative;
-  font-size: @font-size-small;
-  margin-top: @spacing-standard;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  width: 100%;
-  display: inline-block;
-  border-radius: @border-radius-base;
-  border: 1px solid @border-color-divider;
-
-  .unsubscribe-header {
-    border-bottom: 1px solid @border-color-divider;
-    padding: 10px;
-    overflow: hidden;
-    text-overflow: ellipsis;
-
-    img {
-      margin-right: 20px;
-    }
-
-    .unsubscribe-title-text {
-      color: @text-color-very-subtle;
-    }
-
-    .unsubscribe-title {
-      color: @text-color;
-    }
-
-    .unsubscribe-button {
-      float: right;
-      border-radius: @border-radius-base;
-      border: 1px solid @border-color-divider;
-      padding: 2px 4px;
-      line-height: 16px;
-
-      &:hover {
-        background: mix(@background-primary, @text-color, 95%);
-      }
-      &:active {
-        background: mix(@background-primary, @text-color, 92%);
-      }
-
-      img {
-        background: @text-color-subtle;
-        margin: 0;
-      }
-    }
-  }
-
-  .unsubscribe-body {
-    padding: @padding-small-horizontal;
-    color: @text-color;
-
-    .unsubscribe-action {
-      display: flex;
-
-      .btn-unsubscribe {
-        margin-right: 10.5px;
-        margin-top: 5px;
-        min-width: 85px;
-        margin: auto;
-
-        &:hover {
-          background: mix(@background-primary, @text-color, 95%);
-        }
-        &:active {
-          background: mix(@background-primary, @text-color, 92%);
-        }
-
-        &.UNSUBSCRIBED {
-          background: @color-success;
-          color: @white;
-        }
-      }
-    }
-  }
-}
-
-body.platform-win32 {
-  .unsubscribe-wrapper {
-    border-radius: 0;
-    .unsubscribe-body {
-      .unsubscribe-action {
-        .btn-rsvp {
-          border-radius: 0;
-        }
-      }
-    }
-  }
-}
