+++ app/lang/de.json
@@ -804,6 +804,7 @@
   "Unstarring": "Hervorhebung wird entfernt",
   "Untitled": "Ohne Titel",
   "Untitled Rule": "Ohne Titel",
+  "Unsubscribe Now": "Jetzt Abmelden",
   "Update Connection Settings...": "Verbindungseinstellungen aktualisieren...",
   "Update Error": "Aktualisierungsfehler",
   "Updates": "Aktualisierungen",
